+++ drivers/flexspi/fsl_flexspi_dma.c
@@ -157,8 +157,8 @@ static status_t FLEXSPI_WriteDataDMA(FLEXSPI_Type *base, flexspi_dma_handle_t *h
     uint32_t bytesPerDes;
     uint8_t desCount;
     uint8_t remains;
-    uint32_t srcInc;
-    uint32_t dstInc;
+    uint8_t srcInc;
+    uint8_t dstInc;
 
     /* Source address interleave size */
     srcInc = kDMA_AddressInterleave1xWidth;
@@ -205,7 +205,7 @@ static status_t FLEXSPI_WriteDataDMA(FLEXSPI_Type *base, flexspi_dma_handle_t *h
     remains     = (uint8_t)(dataSize - (uint32_t)desCount * (uint32_t)handle->nbytes);
     if (remains > 0U)
     {
-        uint32_t width = (uint32_t)kFLEXPSI_DMAnSize1Bytes;
+        uint8_t width = (uint8_t)kFLEXPSI_DMAnSize1Bytes;
         DMA_SetupDescriptor(&s_flexspiDes[desCount - 1U],
                             DMA_CHANNEL_XFER(false, true, true, false, width, srcInc, dstInc, remains),
                             (void *)(uint64_t *)((uint32_t)data + desCount * bytesPerDes), txFifoBase, NULL);
@@ -282,7 +282,7 @@ static status_t FLEXSPI_WriteDataDMA(FLEXSPI_Type *base, flexspi_dma_handle_t *h
     {
         DMA_SetupDescriptor(&s_flexspiDes[i - 1U],
                             DMA_CHANNEL_XFER((nextDesc == NULL) ? false : true, true, (nextDesc == NULL) ? true : false,
-                                             false, (uint32_t)handle->nsize, srcInc, dstInc, remains),
+                                             false, (uint8_t)handle->nsize, srcInc, dstInc, remains),
                             (void *)(uint64_t *)((uint32_t)data + i * bytesPerDes), txFifoBase, nextDesc);
         nextDesc = &s_flexspiDes[i - 1U];
     }
@@ -290,7 +290,7 @@ static status_t FLEXSPI_WriteDataDMA(FLEXSPI_Type *base, flexspi_dma_handle_t *h
     DMA_PrepareChannelTransfer(
         &txChannelConfig, (void *)data, txFifoBase,
         DMA_CHANNEL_XFER((nextDesc == NULL) ? false : true, true, (nextDesc == NULL) ? true : false, false,
-                         (uint32_t)handle->nsize, srcInc, dstInc, bytesPerDes),
+                         (uint8_t)handle->nsize, srcInc, dstInc, bytesPerDes),
         kDMA_MemoryToMemory, &dmaTxTriggerConfig, nextDesc);
 
     (void)DMA_SubmitChannelTransfer(handle->txDmaHandle, &txChannelConfig);
@@ -316,8 +316,8 @@ static status_t FLEXSPI_ReadDataDMA(FLEXSPI_Type *base, flexspi_dma_handle_t *ha
     uint32_t bytesPerDes;
     uint8_t remains;
     uint8_t desCount;
-    uint32_t srcInc;
-    uint32_t dstInc;
+    uint8_t srcInc;
+    uint8_t dstInc;
 
     /* Source address interleave size */
     srcInc = kDMA_AddressInterleave1xWidth;
@@ -358,7 +358,7 @@ static status_t FLEXSPI_ReadDataDMA(FLEXSPI_Type *base, flexspi_dma_handle_t *ha
 
     if (remains > 0U)
     {
-        uint32_t width = (uint32_t)kFLEXPSI_DMAnSize1Bytes;
+        uint8_t width = (uint8_t)kFLEXPSI_DMAnSize1Bytes;
         DMA_SetupDescriptor(&s_flexspiDes[desCount - 1U],
                             DMA_CHANNEL_XFER(false, true, true, false, width, srcInc, dstInc, remains), rxFifoBase,
                             (void *)(uint64_t *)((uint32_t)data + desCount * bytesPerDes), NULL);
@@ -427,7 +427,7 @@ static status_t FLEXSPI_ReadDataDMA(FLEXSPI_Type *base, flexspi_dma_handle_t *ha
     {
         DMA_SetupDescriptor(&s_flexspiDes[i - 1U],
                             DMA_CHANNEL_XFER((nextDesc == NULL) ? false : true, true, (nextDesc == NULL) ? true : false,
-                                             false, (uint32_t)handle->nsize, srcInc, dstInc, remains),
+                                             false, (uint8_t)handle->nsize, srcInc, dstInc, remains),
                             rxFifoBase, (void *)(uint64_t *)((uint32_t)data + i * bytesPerDes), nextDesc);
         nextDesc = &s_flexspiDes[i - 1U];
     }
@@ -435,7 +435,7 @@ static status_t FLEXSPI_ReadDataDMA(FLEXSPI_Type *base, flexspi_dma_handle_t *ha
     DMA_PrepareChannelTransfer(
         &rxChannelConfig, rxFifoBase, (void *)data,
         DMA_CHANNEL_XFER((nextDesc == NULL) ? false : true, true, (nextDesc == NULL) ? true : false, false,
-                         (uint32_t)handle->nsize, srcInc, dstInc, bytesPerDes),
+                         (uint8_t)handle->nsize, srcInc, dstInc, bytesPerDes),
         kDMA_MemoryToMemory, &dmaRxTriggerConfig, nextDesc);
 
     (void)DMA_SubmitChannelTransfer(handle->rxDmaHandle, &rxChannelConfig);
